perceptron/digit_reader/requirements.txt
fastapi
uvicorn
scikit-learn
pillow
numpy
jinja2
python-multipart
